Exclude Nuxt package
This is a temporary 'fix' to exclude the unfinished Nuxt package.

Further commits will completely move all Nuxt related code to the intended 'nuxt' feature branch.
This is not super ideal, as the code can still be found in the history,
but this enables us to to clean merge when the Nuxt package is finished.